+++ cirneco-reflection-utils/src/main/java/it/ozimov/cirneco/reflection/NoArgumentConstructor.java
@@ -8,9 +8,11 @@ import java.lang.annotation.Target;
 /**
  * Annotation for default constructors. A default constructor is a constructor with no
  * arguments.
+ *
+ * @since version 0.6.0 for JDK7
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.CONSTRUCTOR)
-public @interface DefaultConstructor {
+public @interface NoArgumentConstructor {
 
 }