+++ cirneco-reflection-utils/src/main/java/it/ozimov/cirneco/reflection/ScanNoArgumentConstructor.java
@@ -14,8 +14,6 @@ import org.reflections.util.FilterBuilder;
 
 import javax.annotation.Nonnull;
 import java.lang.reflect.Constructor;
-import java.util.Arrays;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
 
@@ -25,16 +23,21 @@ import static com.google.common.base.Preconditions.checkState;
 import static java.lang.Character.isJavaIdentifierPart;
 import static java.lang.Character.isJavaIdentifierStart;
 
-public class ScanDefaultConstructor {
+/**
+ * Scan for a no-argument constructor
+ *
+ * @since version 0.6.0 for JDK7
+ */
+public class ScanNoArgumentConstructor {
 
     private static final Set<String> RESERVED_KEYWORDS = initReservedKeywords();
 
-    @DefaultConstructor
-    private ScanDefaultConstructor() {
+    @NoArgumentConstructor
+    private ScanNoArgumentConstructor() {
 
     }
 
-    public static Optional<Constructor> getDefaultConstructor(@Nonnull final Class clazz){
+    public static Optional<Constructor> getNoArgumentConstructor(@Nonnull final Class clazz) {
         checkNotNull(clazz);
 
         final List<ClassLoader> classLoadersList = ImmutableList.of(
@@ -48,22 +51,21 @@ public class ScanDefaultConstructor {
                 .filterInputsBy(new FilterBuilder().includePackage(clazz)));
 
 
-        final Set<Constructor> constructors = reflections.getConstructorsAnnotatedWith(DefaultConstructor.class);
+        final Set<Constructor> constructors = reflections.getConstructorsAnnotatedWith(NoArgumentConstructor.class);
         checkState(constructors.size() <= 1, "The class cannot have multiple default constructors, while %d found",
                 constructors.size());
 
-        if(constructors.isEmpty()){
+        if (constructors.isEmpty()) {
             return Optional.absent();
-        }
-        else{
+        } else {
             final Constructor constructor = constructors.iterator().next();
             final Class[] parameterTypes = constructor.getParameterTypes();
-            checkState(parameterTypes.length==0, "Constructor '%s' has input parameters, expected none", constructor.toGenericString());
+            checkState(parameterTypes.length == 0, "Constructor '%s' has input parameters, expected none", constructor.toGenericString());
             return Optional.of(constructor);
         }
     }
 
-    public static Set<Constructor> getAllDefaultConstructors(@Nonnull final String aPackage) {
+    public static Set<Constructor> getAllNoArgumentConstructors(@Nonnull final String aPackage) {
         checkNotNull(aPackage);
         checkArgument(isValidClassIdentifier(aPackage),
                 "Given package '%s' is not a valid java identifier", aPackage);
@@ -79,10 +81,10 @@ public class ScanDefaultConstructor {
                 .setUrls(ClasspathHelper.forClassLoader(classLoadersList.toArray(new ClassLoader[0])))
                 .filterInputsBy(new FilterBuilder().include(FilterBuilder.prefix(aPackage))));
 
-        final Set<Constructor> constructors = reflections.getConstructorsAnnotatedWith(DefaultConstructor.class);
-        for(final Constructor constructor : constructors){
+        final Set<Constructor> constructors = reflections.getConstructorsAnnotatedWith(NoArgumentConstructor.class);
+        for (final Constructor constructor : constructors) {
             final Class[] parameterTypes = constructor.getParameterTypes();
-            checkState(parameterTypes.length==0, "Constructor '%s' has input parameters, expected none", constructor.toGenericString());
+            checkState(parameterTypes.length == 0, "Constructor '%s' has input parameters, expected none", constructor.toGenericString());
         }
         return constructors;
     }
@@ -93,7 +95,7 @@ public class ScanDefaultConstructor {
             if (part.isEmpty()) {
                 return false;
             }
-            if (RESERVED_KEYWORDS.contains(part)){
+            if (RESERVED_KEYWORDS.contains(part)) {
                 return false;
             }
             if (!isJavaIdentifierStart(part.codePointAt(0))) {