+++ cirneco-reflection-utils/src/test/java/testing/reflection/clazz/has/no/annotations/TestClass.java
@@ -2,7 +2,7 @@ package testing.reflection.clazz.has.no.annotations;
 
 public class TestClass {
 
-    private TestClass(){
+    private TestClass() {
 
     }