+++ cirneco-reflection-utils/src/test/java/testing/reflection/clazz/with/constructor/with/params/TestClass.java
@@ -1,11 +1,11 @@
 package testing.reflection.clazz.with.constructor.with.params;
 
-import it.ozimov.cirneco.reflection.DefaultConstructor;
+import it.ozimov.cirneco.reflection.NoArgumentConstructor;
 
 public class TestClass {
 
-    @DefaultConstructor
-    private TestClass(String param){
+    @NoArgumentConstructor
+    private TestClass(String param) {
 
     }
 }