+++ hamcrest-matchers/java7-hamcrest-matchers/src/main/java/it/ozimov/cirneco/hamcrest/java7/date/IsDate.java
@@ -11,7 +11,7 @@ import java.util.Date;
 import static com.google.common.base.Preconditions.checkArgument;
 
 /**
- * it.ozimov.cirneco.hamcrest.java7.Is {@linkplain Date} in a given week day?
+ * Is {@linkplain Date} in a given week day?
  *
  * @since version 0.1 for JDK7
  */